Rename and update the documentation
`JMBaseException.java` -> `JMatrixBaseException.java`
